requirements.txt
aiogram==3.7.0
yt-dlp==2024.4.9
python-dotenv==1.0.0
